Fixing Visual C++ warning about size_t to int conversion
We have no choice, since V8 String constructor accepts int for a string length

+++ v8pp/to_v8.hpp
@@ -130,7 +130,7 @@ inline v8::Handle<T> to_v8(v8::Handle<T> src)
 
 inline v8::Handle<v8::Value> to_v8(std::string const& src)
 {
-	return v8::String::New(src.data(), src.length());
+	return v8::String::New(src.data(), (int)src.length());
 }
 
 inline v8::Handle<v8::Value> to_v8(char const *src)
@@ -143,7 +143,7 @@ static_assert(sizeof(wchar_t) == sizeof(uint16_t), "wchar_t has 16 bits");
 
 inline v8::Handle<v8::Value> to_v8(std::wstring const& src)
 {
-	return v8::String::New((uint16_t const*)src.data(), src.length());
+	return v8::String::New((uint16_t const*)src.data(), (int)src.length());
 }
 
 inline v8::Handle<v8::Value> to_v8(wchar_t const *src)
